Fix the new theme's atrocious padding around labels
It seems as though the update has also caused a few bugs (labels not updating after changing them, live preview not working, etc.), so I might fix those if/when I get around to it.

It should be noted that this is the first update since the new theme was introduced. If you are still using the old theme, newer versions of this extension may not work as intended. It is recommended to either stick with an older version of the extension, or just use the new themes.